requirements.txt
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.13.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
gutils-python @ git+https://github.com/navhits/gutils-python.git@8df6d5917905f4da4eadf5ca34fdd297f45dc0fa
httplib2==0.20.4
idna==3.4
numpy==1.23.4
oauthlib==3.2.2
pandas==1.5.1
protobuf==4.21.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
